docproc: do not pass AM_CFLAGS to the compiler
AM_CFLAGS is used to pass compiler flags detected by configure script
for the target compiler. As the host compiler might not support the
selected features, it is an error to pass these extra options to the
docproc compiler.
As this is a simple application used for building documentation, the
extra protection added by advanced compiler features isn't required
anyway, so drop it completely.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@imgtec.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>