Add multiple operations infrastructure
The AF_ALG interface uses two file descriptors: tfmfd and opfd. The opfd
references multiple operations on the tfm (comparable to the request
data structures in the kernel crypto API). A user may perform multiple
parallel operations on one tfm.

This patch set introduces the framework to support multiple operations
on one tfm file descriptor. This logic is orthogonal to AIO support.

Currently, this infrastructure is unused and is effectively a no-op for
all supported cipher types. A potential future support for the KPP API
would use that framework for supporting the following concept: a private
key is generated in the kernel and retained in the kernel -- only the
public key is exported. A following shared secret request provides the
public key of the remote entity to the kernel. This way, the kernel
supports a full DH operation where the private key is protected and yet
a shared secret can be generated.

Signed-off-by: Stephan Mueller <smueller@chronox.de>