add large encryption test
This test is inteded to verify the proper chaining of TX SGLs in the
kernel. One TX SGL has as many entries as its meta data fits into one
page. When more operation is performed, multiple TX SGLs are chained.
This test creates as many operations to enforce that at least two TX
SGLs are chained.

Signed-off-by: Stephan Mueller <smueller@chronox.de>