remove the specific retain API
When the caler wants the kernel to generate the private DH/ECDH key, all
it needs to do is to invoke a setkey operation with an empty private key
buffer. If the kernel does not support this operation (whih is the case
for DH at the time of writing), -EOPNOTSUPP is returned.

Signed-off-by: Stephan Mueller <smueller@chronox.de>