test: Allow running tests outside of build environment
This patch slightly refactors the test scripts so they can be installed
along with the binaries and ran on the installed system. The old way of
running test-invocation.sh in the build environment still works.

The test scripts are installed into %pkglibexecdir%, along with the test
binaries (so that the test scripts can easily find them). Note that the
test binaries are also still installed in %bindir% to not break the
backwards compatibility.

Distribuitons are recommended to put the test files into a separate
subpackage that can be installed for debugging/QA purposes.