add kcapi_cipher_stream_update_last
The kernel 5.5, the kernel crypto API received an update to the
symmetric cipher AF_ALG interface code that is visible for stream
ciphers using an underlying block cipher.

When sending data, the kernel must know whether the last data is
received. Only when the kernel is informed about the fact that the last
data is received, the kernel will return all data including the last
block that may not be a multiple of the block cipher.

To inform the kernel about the last data, the new API call of
kcapi_cipher_stream_update_last is added. It is permissible to invoke
the API call with a NULL buffer.

For example, if ctr(aes) is used to encrypt 257 bytes of data, the
kernel would return only 256 bytes of ciphertext without
having the call kcapi_cipher_stream_update_last was invoked. Only when
this API call was invoked, the last byte is returned.

Signed-off-by: Stephan Mueller <smueller@chronox.de>