Disable AIO parallel tests
Parallel AIO requests on the same cipher handle are undefined because
the driver may choose to either serialize the request or perform
parallel operations with separate states. Only for non-AIO drivers
(e.g. the C implementations or AESNI) that are invoked via the AIO
interface, the requests will be serialized.

Due to this undefined behavior of AIO, I disabled the test for parallel AIO.

Reported-by: Ondrej Mosnáček
Signed-off-by: Stephan Mueller <smueller@chronox.de>