Streamline size of MMAP segment
Testing of Intel Ivybridge and an Armv5 with slow drives shows that
mapping file does not increase hashing beyond 4M segmented mapping.
Recent ssd drives have load with speeds over 1G which amounts to
256Hz 4M loads.The kernel timer runs at 100Hz which is not enough
to account for 256Hz=1G/4M For such drives 16M mapping is minimum.
This appears to be true in case buffers are managed by workqueues.

Signed-off-by: Brandur Simonsen <brandur.simonsen@gmail.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>