kcapi-hasher: Fix double munmap in hasher()
In hasher(), it is possible that munmap(2) gets called twice on the same
range. This was reported as a bug by Coverity scan. In this case it
shouldn't matter, since there should be no mmap() calls inbetween, but
it's better to avoid the double call anyway.

Fix it by moving the second munmap() from below the 'out' label to the
only error path where it is actually needed and replace the gotos with
direct return. This makes the code easier to follow since the mmmap()
and munmap() calls are close together.

Fixes: fc4a9d8ca95c ("suppot files >4GB")
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>