reenable parallel AIO requests
Parallel AIO requests are undefined - it may be the case that such
parallel requests are serialized by the driver or that they are
processed independently of each other. Yet, we perform testing to
verify that the kernel code gracefully handles that scenario.

The cipher implementation can only return one of two results: either a
CBC block chaining is applied to the two input blocks, or both blocks
are encrypted separately.

Signed-off-by: Stephan Mueller <smueller@chronox.de>