Use sendmsg as fallback for vmsplice in all occasions
The vmsplice system call is commonly the preferred method for
communicating with the kernel. However, this system call has the
following limitations:

1. At most only 16 pages are allowed to be sent to the kernel -
   this is a hard limit in the vmsplice system call.
2. At most, the entire buffer size must not be larger than the
   buffer size of the pipe used for the splice. If desired, this
   buffer size can be enlarged with kcapi_set_maxsplicesize().
   On the other side, the maximum buffer size can be obtained via
   kcapi_get_maxsplicesize().
3. vmsplice cannot be used for a zero buffer.

If any of the aforementioned conditions is not satisfied, the sendmsg
system call is used.

In order for this to work, the pipe buffer size is increased by one more
page as seemingly the kernel and/or the libkcapi operation occupies
parts of the pipe buffer.

In addition, the library gained the calls for querying and setting the
maximum buffer size.

Signed-off-by: Stephan Mueller <smueller@chronox.de>