Use size_t / ssize_t
The library uses the sendmsg / recv / splice / vmsplice system calls for
the communication with the kernel. Those are defined to use
size_t/ssize_t variable types. The library on the other hand used
uint32_t / int32_t which does not comply and caused a lot of warnings
with -Wconversion.

This patch started by changing the variable types used to invoke the
system call to ssize_t / size_t. This change metastized into all code
parts of the library to use the updated data types.

The change now allows the enabling of -Wconversion and limit the number
of type casts.

Signed-off-by: Stephan Mueller <smueller@chronox.de>