Version symbols to maintain ABI compatibility
Before 1.3.0 many functions had a different signature.
Add back compatibility functions and actually use symbol versioning to
maintain binary compatibility.

Note that the 1.3.0 release is just broken any build against that
version will break any previous or later built binaries that link to it,
and there is nothing that can be done to fix that version.

Signed-off-by: Simo Sorce <simo@redhat.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>