Drop use of visibility=hidden
The version-script file already controls what is exported so this flag is
basically unnecessary now. It requires to sprinkle a bunch of attributes on
all exported function via the DSO_PUBLIC and other SYMVER macros.

Removing it makes it simpler to deal with exported symbols, leaving
control only to the version.lds file.

Signed-off-by: Simo Sorce <simo@redhat.com>