Add Support for the kcapi_rng_setentropy API
The kernel config option CONFIG_CRYPTO_USER_API_RNG_CAVP was
introduced to the linux kernel in version 5.10 and can be used
to perform CAVP testing on the kernel DRBGs. This allows the
DRBGs to be tested from user-space and therefore the ability
to set the initial entropy of the DRBG could be a useful addition
to this library.

This change:

* Adds support for the kcapi_rng_setentropy API
* Adds the required logic in kernel-if to use the socket option
  to set the DRBG's entropy.
* Changes the values for ALG_SET_PUBKEY, et. al. to accomodate
  the ALG_SET_DRBG_ENTROPY socket option
* Adds a private _kcapi_common_setentropy API

Signed-off-by: Puru Kulkarni <puruk@protonmail.com>