Add Support for the kcapi_rng_setaddtl API
The Kernel's documentation suggests that in order to provide
additional data when CAVP testing for the kernel DRBG is enabled,
one must use the sendmsg() API in order to pass in the required
additional data.

Therefore, we must provide a way for libkcapi, to call
send/sendmsg() with an initialized and seeded RNG handle so that
additional data can be provided and consumed on the call to
kcapi_rng_generate().

This change:

* Adds the kcapi_rng_setaddtl API
* Adds this API to version.lds

Signed-off-by: Puru Kulkarni <puruk@protonmail.com>