kcapi-hasher ship the multi-call binary
Ship the kcapi-hasher multi-call binary, and install symbolic relative
symlinks to it in the libexec directory. This is packaging policy
compliant with all major Linux distributions (they shun hard
links). And this allows for libkcapi tools to co-exist and be
co-installed with coreutils / rust-coreutils / openssl / etc
implementations. Installing kcapi-hasher utility directly makes it
more accessible and more useful in many context when one wants to
compare and examine kcapi-hasher implementation with other ones.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>