Transition all self-checks to sha512hmac
Now that all applications are symlinks, they cannot have different
self-checks, as they all try to read the same .kcapi-hasher.hmac. This
is thus a breaking change as libkcapi.so and most *sums apps will now
now perform self-checks using hmaccalc.

If this is not acceptable and previous scheme of using fipscheck for
self-check of libkcapi.so fipscheck fipshmac is required fipsselfcheck
function will need adjustment to read and parse /proc/self/cmdline
instead of /proc/self/exe.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>