remove SPLICE_F_GIFT
Passing SPLICE_F_GIFT to vmsplice() allows the kernel to excise the
entire page(s) containing the data from the calling process's address
space (typically when page is spliced out of the pipe with SPLICE_F_MOVE).
If this page is in the heap, this could wreck malloc's metadata there.

SPLICE_F_GIFT does not gain anything when just passing the data to
AF_ALG. The pages from the userspace process are spliced into the pipe
anyway, even without SPLICE_F_GIFT. At present, it merely sets a flag
that the other end can take note of.

Reported-by: David Howells
Signed-off-by: Stephan Mueller <smueller@chronox.de>